+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+- Fixed `Console.export_svg()` generating a dangling `clip-path` reference for the last line by emitting a `<clipPath>` definition for every emitted line https://github.com/Textualize/rich/issues/4215
+
 ## [15.0.0] - 2026-04-12
 
 ### Changed